+++ chemprop/v2/cli/utils/utils.py
@@ -1,6 +1,6 @@
 from typing import Any, Type
 
-from chemprop.v2.models import loss
+from chemprop.v2.nn import loss
 from chemprop.v2.models.modules import readout
 
 __all__ = ["pop_attr"]
