+++ chemprop/v2/models/model.py
@@ -9,9 +9,9 @@ from torch import nn, Tensor, optim
 from chemprop.v2.data.dataloader import TrainingBatch
 from chemprop.v2.featurizers.molgraph import BatchMolGraph
 from chemprop.v2.nn import MessagePassingBlock, Aggregation, Readout
-from chemprop.v2.models.loss import LossFunction
-from chemprop.v2.models.metrics import Metric
-from chemprop.v2.models.schedulers import NoamLR
+from chemprop.v2.nn.loss import LossFunction
+from chemprop.v2.nn.metrics import Metric
+from chemprop.v2.schedulers import NoamLR
 
 
 class MPNN(pl.LightningModule):
