+++ chemprop/v2/nn/agg.py
@@ -5,7 +5,7 @@ import torch
 from torch import Tensor, nn
 
 from chemprop.v2.utils import ClassRegistry
-from chemprop.v2.models.hparams import HasHParams
+from chemprop.v2.utils.hparams import HasHParams
 
 AggregationRegistry = ClassRegistry()
