+++ chemprop/v2/nn/message_passing/base.py
@@ -3,7 +3,7 @@ from typing import Protocol
 from torch import nn, Tensor
 
 from chemprop.v2.featurizers.molgraph import BatchMolGraph
-from chemprop.v2.models.hparams import HasHParams
+from chemprop.v2.utils.hparams import HasHParams
 
 
 class MessagePassingProto(Protocol):
