+++ chemprop/v2/nn/message_passing/molecule.py
@@ -7,7 +7,7 @@ from torch import Tensor, nn
 from chemprop.v2.conf import DEFAULT_ATOM_FDIM, DEFAULT_BOND_FDIM, DEFAULT_HIDDEN_DIM
 from chemprop.v2.exceptions import InvalidShapeError
 from chemprop.v2.featurizers import BatchMolGraph
-from chemprop.v2.models.utils import Activation, get_activation_function
+from chemprop.v2.nn.utils import Activation, get_activation_function
 from chemprop.v2.models.modules.message_passing.base import MessagePassingBlock
