+++ chemprop/v2/utils/__init__.py
@@ -1,3 +1,4 @@
+from .hparams import HasHParams
 from .mixins import ReprMixin
 from .registry import ClassRegistry
 from .utils import *
